requirements.txt
streamlit
pandas
fpdf
google-generativeai
pytest
